Changes to data_manipolation file
Sensor_data was modified into features, which isn't the
script that has to do so. Now sensor_data is zipped inside
data_manipolation.
There is also a changing in which data we store into the database, from
normalized_data to sensor_data. This will be useful for frequency
analysis also.